Align Active requirements with intent
Reaching active status is simply meant as a checkpoint that the working
group satisfied the incubation tasks and is operating correctly. Clarify
and adjust the Active requirements to a reasonable minimum amount of
time to demonstrate that incubation was completed in good faith and
the working group is operating in a healthy and repeatable manner.

Signed-off-by: Dan Middleton <dan.middleton@intel.com>