Adding TAC and SCIR election process description
The procedure for the TAC and SCIR elections has not yet been officially
documented. In preparation for the 2023 elections of the TAC and SCIR members,
the Election Officials would like to document the entire election process,
including the specific timeline for the 2023 elections.

This document is based on the procedure used during the 2022 elections [1], based on
discussions in the TAC in preparation for the 2022 election [2].

[1] https://lists.openssf.org/g/openssf-tac/message/406
[2] https://lists.openssf.org/g/openssf-tac/message/364

Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>