clarify terms of appointed and elected members
Given there's been discussion of whether the charter perhaps evolves
to stagger terms and the election/appointing cadence is still early
in its operation, it is useful to be able to see which members came
into the board at what time.  Perhaps eventually this shows via the
git history, but that too takes time evolve and settle and is less
transparent for a typical reader.

Signed-off-by: Tim Pepper <tpepper@vmware.com>