Update documentation on project creation process
The documented process to create a project
doesn't seem to match what is actually done, so I've modified
the text to describe what I think is *actually* done.

TAC, do you agree? If not, what should it say instead?
I'm *not* a TAC member, so I do not decide this, but I thought it
might be easier to have a pull request to review.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>