Add build a community
This solves #244. This is a document for helping people and
organizations that have open sourc projects they would like to
eventually contribute to the OpenSSF but don't meet the
requirements to be a sandbox project. In particular the need for
maintainers from 2 organizations can be difficult when the project
was started by an individual or internally at an organization.

This document highlights some simple practices to get started with
building a new community.

Signed-off-by: Michael Lieberman <mlieberman85@gmail.com>