Update funding_application.yml to reference process doc
This PR removes any reference to the process so that we have a single source of truth in the process documentation. 

The issue template was made before the process documentation but now that that is in place the issue does not need the information and it is best to reference the other location. 

Signed-off-by: Amanda L Martin <hythloda@gmail.com>