Iterate on funding application
Arguably, our second funding cycle was a success with at least two
requests approved! When we went to evaluate the applications, it became
clear that an exact amount was helpful for budgeting, and it wasn't
clear that funding requests should come from a specific OpenSSF TI.
These proposed changes aim to make the funding process more clear to
applicants.

Signed-off-by: Zach Steindler <steiza@github.com>