Create sandbox project for the model signing work
We have a working group that meets for model signing work, as part of
https://github.com/ossf/ai-ml-security/issues/10. Since this working
group helps in developing https://github.com/sigstore/model-transparency
and building standards around it, it needs to be officially a project,
not a WG.

We add the project at a sandbox stage.

Please note that there are 2 repos invovled here:

- https://github.com/sigstore/model-transparency which will be owned by
  Sigstore and is just the implementation work for the library for model
  signing
- a new repository to be created under https://github.com/ossf to
  represent standard documents, as outputs of this project

These two repositories should work in unison to achieve a common goal.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>