Update tac-member-R&Rs.md
As Crob and I were preparing for an update to the OpenSSF MVSR going into 2025, we found a list of TAC responsibilities. Adds here reflect some of what was in that document, as well as a few things I've been doing that I've seen us collectively doing that think should be communicated out prior to the election.

Signed-off-by: Sarah Evans <140007827+sevansdell@users.noreply.github.com>